CMainFrm: ditched CSplitterWindow in favor of a manual splitter. This improves performance and gets rid of the tab control background drawing bugs. CNzbView/CFileView: fixed progress bar drawing for Windows XP. Still doesn't work when theming is disabled. removed CDiskWriter singleton, access via CNewzflow::Instance()->diskWriter now. added CPostProcessor, calls par2.exe (tbb-multicore version used for testing), and parses output for verification. CPostProcessor::Par2Verify still work-in-progress. Need to clean up pipes, etc. removed atldbgmem.h because it doesn't work with STL.

+++ ConnectionView.cpp
@@ -37,7 +37,7 @@ LRESULT CConnectionView::OnTimer(UINT /*uMsg*/, WPARAM /*wParam*/, LPARAM /*lPar
 			CDownloader* dl = CNewzflow::Instance()->downloaders[i];
 			CString s;
 			s.Format(_T("%d"), i+1);
-			if(i >= lvCount) {
+			if((int)i >= lvCount) {
 				AddItem(i, 0, s);
 			}
 			s.Format(_T("%.1f kB/s"), dl->sock.speed.Get() / 1024.f);
@@ -49,7 +49,7 @@ LRESULT CConnectionView::OnTimer(UINT /*uMsg*/, WPARAM /*wParam*/, LPARAM /*lPar
 			AddItem(i, 2, s);
 		}
 		SetRedraw(FALSE);
-		while(count < lvCount) {
+		while((int)count < lvCount) {
 			DeleteItem(count);
 			lvCount--;
 		}
